Also make top level ruleset settings visible on sidebar
