+++ osu.Game/Overlays/FirstRunSetup/ScreenBehaviour.cs
@@ -83,7 +83,6 @@ namespace osu.Game.Overlays.FirstRunSetup
                         // InputSection is intentionally omitted for now due to its sub-panel being a pain to set up.
                         new UserInterfaceSection(),
                         new GameplaySection(),
-                        new RulesetSection(),
                         new AudioSection(),
                         new GraphicsSection(),
                         new OnlineSection(),
