+++ osu.Game/Overlays/Settings/Sections/RulesetSection.cs
@@ -1,44 +1,27 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Linq;
-using osu.Framework.Allocation;
 using osu.Framework.Graphics;
-using osu.Framework.Graphics.Sprites;
 using osu.Framework.Localisation;
-using osu.Game.Graphics;
-using osu.Game.Localisation;
 using osu.Game.Rulesets;
 
 namespace osu.Game.Overlays.Settings.Sections
 {
     public partial class RulesetSection : SettingsSection
     {
-        public override LocalisableString Header => RulesetSettingsStrings.Rulesets;
+        private readonly Ruleset ruleset;
 
-        public override Drawable CreateIcon() => new SpriteIcon
+        public RulesetSection(Ruleset ruleset, SettingsSubsection section)
         {
-            Icon = OsuIcon.Rulesets
-        };
+            UseSmallerSidebarButton = true;
 
-        [BackgroundDependencyLoader]
-        private void load(RulesetStore rulesets)
-        {
-            foreach (Ruleset ruleset in rulesets.AvailableRulesets.Select(info => info.CreateInstance()))
-            {
-                try
-                {
-                    SettingsSubsection? section = ruleset.CreateSettings();
+            this.ruleset = ruleset;
 
-                    if (section != null)
-                        Add(section);
-                }
-                catch (Exception e)
-                {
-                    RulesetStore.LogRulesetFailure(ruleset.RulesetInfo, e);
-                }
-            }
+            Add(section);
         }
+
+        public override LocalisableString Header => ruleset.RulesetInfo.Name;
+
+        public override Drawable CreateIcon() => ruleset.CreateIcon();
     }
 }
