+++ osu.Game/Overlays/SettingsOverlay.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-#nullable disable
-
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,6 +17,7 @@ using osu.Game.Localisation;
 using osu.Game.Overlays.Settings;
 using osu.Game.Overlays.Settings.Sections;
 using osu.Game.Overlays.Settings.Sections.Input;
+using osu.Game.Rulesets;
 using osuTK.Graphics;
 
 namespace osu.Game.Overlays
@@ -29,34 +28,9 @@ namespace osu.Game.Overlays
         public LocalisableString Title => SettingsStrings.HeaderTitle;
         public LocalisableString Description => SettingsStrings.HeaderDescription;
 
-        protected override IEnumerable<SettingsSection> CreateSections()
-        {
-            var sections = new List<SettingsSection>
-            {
-                // This list should be kept in sync with ScreenBehaviour.
-                new GeneralSection(),
-                new SkinSection(),
-                new InputSection(createSubPanel(new KeyBindingPanel())),
-                new UserInterfaceSection(),
-                new GameplaySection(),
-                new RulesetSection(),
-                new AudioSection(),
-                new GraphicsSection(),
-                new OnlineSection(),
-                new MaintenanceSection(),
-                new DebugSection()
-            };
-
-            var today = DateTimeOffset.Now;
-            if (today.Month == 4 && today.Day == 1)
-                sections.Insert(9, new AfToggleSection());
-
-            return sections;
-        }
-
         private readonly List<SettingsSubPanel> subPanels = new List<SettingsSubPanel>();
 
-        private SettingsSubPanel lastOpenedSubPanel;
+        private SettingsSubPanel? lastOpenedSubPanel;
 
         protected override Drawable CreateHeader() => new SettingsHeader(Title, Description);
 
@@ -74,8 +48,7 @@ namespace osu.Game.Overlays
 
         public override bool AcceptsFocus => lastOpenedSubPanel == null || lastOpenedSubPanel.State.Value == Visibility.Hidden;
 
-        public void ShowAtControl<T>()
-            where T : Drawable
+        public void ShowAtControl<T>() where T : Drawable
         {
             // if search isn't cleared then the target control won't be visible if it doesn't match the query
             SearchTextBox.Current.SetDefault();
@@ -90,6 +63,56 @@ namespace osu.Game.Overlays
             }
 
             SectionsContainer.ScrollTo(SectionsContainer.ChildrenOfType<T>().Single());
+        }
+
+        protected override float ExpandedPosition => lastOpenedSubPanel?.State.Value == Visibility.Visible ? -PANEL_WIDTH : base.ExpandedPosition;
+
+        [BackgroundDependencyLoader]
+        private void load(RulesetStore rulesets)
+        {
+            var sections = new List<SettingsSection>
+            {
+                // This list should be kept in sync with ScreenBehaviour.
+                new GeneralSection(),
+                new SkinSection(),
+                new InputSection(createSubPanel(new KeyBindingPanel())),
+                new UserInterfaceSection(),
+                new GameplaySection()
+            };
+
+            foreach (Ruleset ruleset in rulesets.AvailableRulesets.Select(info => info.CreateInstance()))
+            {
+                try
+                {
+                    SettingsSubsection? section = ruleset.CreateSettings();
+
+                    if (section != null)
+                        sections.Add(new RulesetSection(ruleset, section));
+                }
+                catch (Exception e)
+                {
+                    RulesetStore.LogRulesetFailure(ruleset.RulesetInfo, e);
+                }
+            }
+
+            sections.AddRange(new SettingsSection[]
+            {
+                new AudioSection(),
+                new GraphicsSection(),
+                new OnlineSection(),
+                new MaintenanceSection(),
+                new DebugSection()
+            });
+
+            var today = DateTimeOffset.Now;
+            if (today.Month == 4 && today.Day == 1)
+                sections.Insert(9, new AfToggleSection());
+
+            foreach (var s in sections)
+                AddSection(s);
+
+            foreach (var s in subPanels)
+                ContentContainer.Add(s);
         }
 
         private T createSubPanel<T>(T subPanel)
@@ -127,15 +150,6 @@ namespace osu.Game.Overlays
                     ContentContainer.MoveToX(0, 500, Easing.OutQuint);
                     break;
             }
-        }
-
-        protected override float ExpandedPosition => lastOpenedSubPanel?.State.Value == Visibility.Visible ? -PANEL_WIDTH : base.ExpandedPosition;
-
-        [BackgroundDependencyLoader]
-        private void load()
-        {
-            foreach (var s in subPanels)
-                ContentContainer.Add(s);
         }
     }
 }
